[sw,tests,pwrmgr] Improve synchronization
Allow each CSR updates that need to transfer to the slow domain to make
synchronization optional.
This enables entry into low power to use a single synchronization when
it is done with the last update.
Update pwrmgr unittests for this change.
Change some tests to use pwrmgr_testutils_enable_low_power.

Signed-off-by: Guillermo Maturana <maturana@google.com>