[lc_ctrl] Beef up internal usage of redundant CSR encoding
The replicated enum encoding of the LC_STATE and TRANSITION_TARGET CSRs
has been added in order to ease hardening on the firmware side.

This patch lists this as an explicit countermeasure, and beefs up the
internal usage of this encoding. I.e., the transition logic only checked
the TRANSITION_TARGET encoding coming from the CSRs before. With this
change, the logic now also makes use of the redundancy present in the
decoded state vector. This is useful in particular when indexing the
transition matrix LUT (to get the valid bits and token indices), since
this indexing operation can now be duplicated.

Signed-off-by: Michael Schaffner <msf@opentitan.org>