[aes] Forward fatal alert to cipher core
Previously, in case of a fatal alert detected outside of the cipher
core, the cipher core would still continue to process the current block
but eventually get blocked by the main FSM not performing the
handshake. Whit this commit, the cipher core will lock up immediately.

This is related to lowRISC/OpenTitan#10422.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>