[otbn,dv] Fix assertion in RND FSM checking
If we are prefetching and a response comes in from the
EDN (edn_rnd_ack_i) at the same time as an instruction reads from
RND (rnd_req_i), we go to the "full" state, not the "reading" state.

The FSM I'd sketched out in the assertions was bogus: when both of
those signals are asserted at the same time, we can't go into both
READING and FULL states!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>