[aes] Update sideload key on second write to the shadowed ctrl reg only
Updating the sideload key on every write to the shadow register can
trigger PRNG reseed operations in the middle of a control register
update complicating software and DV. For this reason, this commit
forwards the phase of the shadow register to the main controller to
only trigger the sideload key update on the second write.

This is related to lowRISC/OpenTitan#10518.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>