[rv_dm dv] Add Hjson spec for debug memory
The comportable rv_dm wrapper only exposes the debug memory as a window.
Our reggen tool infers that as a read-only memory and creates a RAL
model reflecting the same.

In reality, behind the window is a whole slew of CSRs, ROM and SRAM
(program buffer) that are implemented in the vendored-in PULP codebase.

To better aid rv_dm DV, we hand-write that address space with the spec
as reference:
https://github.com/lowRISC/opentitan/blob/master/hw/vendor/pulp_riscv_dbg/doc/debug-system.md#debug-memory

In the subsequent PR, this Hjson will be used for generating a proper
RAL model reflecting the full address map of the debug memory, rather
than just a ROM space inferred from the window.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>