[sw,otbn] Add OTBN code snippet for P-384 Solinas reduce.
Adds a small code snippet demonstrating Solinas modular reduction
algorithm for P-384. This implementation takes 86 cycles for a modular
multiplication and full reduction, or 46 cycles for a modular reduction
alone. It compares favorably to a generic Barrett implementation,
which takes 150 cycles for multiplication and reduction (1.76x speedup)
and 110 for a reduction alone (2.39x speedup).

Signed-off-by: Jade Philipoom <jadep@google.com>