[usbdev] Clean up confusing uses of "differential"
Explicitly refer to a receiver wherever "differential" is used with
"rx".

Previously, the term on RX would mean that an *external* differential
receiver would be used, delivering a *single-ended* data signal to the
IP. However, this top-level definition causes too much confusion.

Similarly, for TX, avoid using "differential" altogether, as neither the
driver nor the signals are truly differential. Instead, refer to the two
interfaces by the component signal names.

Signed-off-by: Alexander Williams <awill@google.com>