[hmac_enc, tests] Introduce the HMAC chip level test (non-irq)
This change uses HMAC example vector from:
https://csrc.nist.gov/CSRC/media/Projects/
Cryptographic-Standards-and-Guidelines/documents/
examples/HMAC_SHA256.pdf

Key length = 100
Tag length = 32

This example vector has key length larger, and not aligned to the
block size. This makes it well suited to check the basic conformance
to the HMAC specification, which states that when key is larger than
the block size, it must be hashed to produce a block sized key.

Please see:
https://csrc.nist.gov/csrc/media/publications/fips/198/archive/
2002-03-06/documents/fips-198a.pdf

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>