[dv,full_chip,clkmgr] Test access to disabled peripheral
Configure a watchdog bite.
Disable the clock to some peripheral and perform a CSR access to it.
The access should cause the CPU to freeze, so the watchdog will bite.
Detect if a reset is due to the watchdog, indicating test success.
Introduce a simple aon_timer testutils function.
Minor cleanup of aon_timer_irq_test.
Minor updates to the full chip testplan.

Signed-off-by: Guillermo Maturana <maturana@google.com>