[otbn,dv] Extend window at end of IMEM error sequence
This is all about an awkward situation where we happen to inject an
IMEM error just after we've fetched the last instruction that we were
going to execute. In this case, there will be no error, which is fine,
but I'd mis-calculated how long you need to wait to figure out whether
it's happened. Fix that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>