[topgen] Remove suffix for mixed port connections
An intermodule signal that connects to an output port and an input on
a module is a mixed port type. These connections produce an internal net
that drives the top-level port, but the requestor module *also* drives
the top-level port (without this change).

Add the "conn_type" attribute to the intermodule signals, so the mixed
connection type is available at the both the port and signal level, and
do not use the "_o" suffix for nets connecting to the requestor module
if the top-level has a mixed connection type.

Signed-off-by: Alexander Williams <awill@google.com>