[dv] Prepare codebase for UVM REG changes
We set the UVM_REG_DATA|ADDR|BYTENABLE_WIDTH values
(compile time macros) based on our 32-bit TL interface.
This works for RAL models that are accessed via the TL
interface.

It unfortunately does not work for RAL models accessed
via the JTAG interface. In principal, these widths should
have not been modified in the first place. We should have
developed our code with the UVM defaults, which assume
a wide enough bus widths (64, 64, 8 respectively).

To support this at this time is difficult (remove these
compile time settings in common_sim_cfg.hjson causes
a widespread beakage.

This PR address SOME of the updates that were needed to
get us to that reality. Slowly, we will fix our codebase
so that we can remove those overrides once and for all.

As for this PR, these changes are needed (as a start)
to support the changes that were made to the jtag_agent,
(subsequent commit) which is instantiated in rv_dm,
lc_ctrl and the full chip testbenches.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>